Update comment in codeql.yml
Changed a comment in the .github/workflows/codeql.yml file to improve clarity. The former comment suggested that this file usually does not need to be changed, but that it should be committed to the repository. The revised comment instructs simply to commit it to the repository, reducing possible confusion.